+++ CHANGES.md
@@ -1,4 +1,5 @@
 #### [unreleased]
+* move list CSS to `class Extras`
 
 #### 5.2.2 / 2018-07-28
 * fix to create CSS when files don't exist